refactor(main): remove old parser and http
