+++ src/agent/mod.rs
@@ -1,3 +1,2 @@
 pub mod http;
-pub mod http2;
 pub mod sql;
